BackEnd/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
boto3==1.38.13
botocore==1.38.13
cachetools==6.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
datasets==3.6.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dspy==2.6.27
dspy-ai==2.6.27
eval_type_backport==0.2.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.3.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.1
json_repair==0.46.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
litellm==1.72.2
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.5
openai==1.79.0
openai-clip==1.0.1
optuna==4.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.1
pyarrow==20.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
s3transfer==0.12.0
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
together==1.5.7
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
